+++ verik-compiler/src/main/kotlin/io/verik/compiler/evaluate/ForEachUnrollTransformerStage.kt
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package io.verik.compiler.transform.upper
+package io.verik.compiler.evaluate
 
-import io.verik.compiler.ast.element.declaration.sv.ESvValueParameter
+import io.verik.compiler.ast.element.declaration.kt.EKtValueParameter
 import io.verik.compiler.ast.element.expression.common.EBlockExpression
 import io.verik.compiler.ast.element.expression.common.ECallExpression
 import io.verik.compiler.ast.element.expression.common.EExpression
@@ -48,7 +48,7 @@ object ForEachUnrollTransformerStage : ProjectStage() {
 
     private class GenerateForBlockReferenceIndexerVisitor : TreeVisitor() {
 
-        val references = HashSet<ESvValueParameter>()
+        val references = HashSet<EKtValueParameter>()
 
         private val innerReferenceIndexerVisitor = InnerReferenceIndexerVisitor()
 
@@ -63,13 +63,13 @@ object ForEachUnrollTransformerStage : ProjectStage() {
 
             override fun visitReferenceExpression(referenceExpression: EReferenceExpression) {
                 val reference = referenceExpression.reference
-                if (reference is ESvValueParameter) references.add(reference)
+                if (reference is EKtValueParameter) references.add(reference)
             }
         }
     }
 
     private class ForEachUnrollTransformerVisitor(
-        private val references: HashSet<ESvValueParameter>
+        private val references: HashSet<EKtValueParameter>
     ) : TreeVisitor() {
 
         private fun getIndices(expression: EExpression): Pair<Int, Int>? {
@@ -123,7 +123,7 @@ object ForEachUnrollTransformerStage : ProjectStage() {
                 }
                 statementGroups.add(statements)
             }
-            val valueParameter = functionLiteralExpression.valueParameters[0].cast<ESvValueParameter>()
+            val valueParameter = functionLiteralExpression.valueParameters[0].cast<EKtValueParameter>()
             statementGroups.forEachIndexed { index, statements ->
                 val valueParameterSubstitutorVisitor = ValueParameterSubstitutorVisitor(
                     valueParameter,
@@ -152,7 +152,7 @@ object ForEachUnrollTransformerStage : ProjectStage() {
         }
 
         private class ValueParameterSubstitutorVisitor(
-            private val valueParameter: ESvValueParameter,
+            private val valueParameter: EKtValueParameter,
             private val index: Int
         ) : TreeVisitor() {