+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Support union types.
 - Modify interface for functions `optional` and `cluster`.
 - Replace generate for blocks with elaborated clusters.
+- Move cluster elaboration to evaluate stage.
 
 ## [0.1.14]
 ### Importer